Stop generating pot file on every test run
